requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.12.0
numpy==2.4.0
openai==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
